+++ pkg/suite/suite.go
@@ -11,15 +11,22 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
+	"github.com/giantswarm/apiextensions-application/api/v1alpha1"
 	applicationv1alpha1 "github.com/giantswarm/apiextensions-application/api/v1alpha1"
 	"github.com/giantswarm/apptest-framework/pkg/client"
 	"github.com/giantswarm/apptest-framework/pkg/config"
 	"github.com/giantswarm/apptest-framework/pkg/state"
+	"github.com/giantswarm/cluster-standup-teardown/pkg/clusterbuilder"
+	"github.com/giantswarm/cluster-standup-teardown/pkg/standup"
+	"github.com/giantswarm/cluster-standup-teardown/pkg/teardown"
 	"github.com/giantswarm/clustertest"
 	"github.com/giantswarm/clustertest/pkg/application"
+	clusterclient "github.com/giantswarm/clustertest/pkg/client"
 	"github.com/giantswarm/clustertest/pkg/logger"
+	"github.com/giantswarm/clustertest/pkg/wait"
 	"k8s.io/apimachinery/pkg/api/errors"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 	cr "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -33,9 +40,9 @@ type suite struct {
 	isUpgrade        bool
 	installNamespace string
 
-	beforeInstall func()
-	beforeUpgrade func()
-	tests         func()
+	afterClusterReady func()
+	beforeUpgrade     func()
+	tests             func()
 }
 
 // New create a new suite instance that allows configuring an App test suite
@@ -74,11 +81,11 @@ func (s *suite) WithValuesFile(valuesFile string) *suite {
 	return s
 }
 
-// BeforeInstall allows configuring tests that will run before the App is installed.
+// AfterClusterReady allows configuring tests that will run as soon as the cluster is up and ready.
 // This allows for running tests to check the current state of the cluster and
 // assert that any pre-requisites are met.
-func (s *suite) BeforeInstall(fn func()) *suite {
-	s.beforeInstall = fn
+func (s *suite) AfterClusterReady(fn func()) *suite {
+	s.afterClusterReady = fn
 	return s
 }
 
@@ -112,15 +119,11 @@ func (s *suite) Run(t *testing.T, suiteName string) {
 
 		mcKubeconfig := os.Getenv("E2E_KUBECONFIG")
 		mcContext := os.Getenv("E2E_KUBECONFIG_CONTEXT")
-		wcName := os.Getenv("E2E_WC_NAME")
-		wcNamespace := os.Getenv("E2E_WC_NAMESPACE")
 		appVersion := os.Getenv("E2E_APP_VERSION")
 
 		// Ensure all require env vars are set
 		Expect(mcKubeconfig).ToNot(BeEmpty(), "`E2E_KUBECONFIG` must be set to the kubeconfig of the test MC")
 		Expect(mcContext).ToNot(BeEmpty(), "`E2E_KUBECONFIG_CONTEXT` must be set to the context to use in the kubeconfig")
-		Expect(wcName).ToNot(BeEmpty(), "`E2E_WC_NAME` must be set to the name of the test WC")
-		Expect(wcNamespace).ToNot(BeEmpty(), "`E2E_WC_NAMESPACE` must be set to namespace the test WC is in")
 		Expect(appVersion).ToNot(BeEmpty(), "`E2E_APP_VERSION` must be set to version of the app to test against")
 
 		state.SetContext(context.Background())
@@ -130,11 +133,75 @@ func (s *suite) Run(t *testing.T, suiteName string) {
 		Expect(err).NotTo(HaveOccurred())
 		state.SetFramework(framework)
 
-		// Setup client for connecting to WC
-		cluster, err := framework.LoadCluster()
+		cb, err := clusterbuilder.GetClusterBuilderForContext(mcContext)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(cb).NotTo(BeNil())
+
+		// Load an existing cluster is env vars are set, otherwise create a new cluster
+		cluster := clusterbuilder.LoadOrBuildCluster(state.GetFramework(), cb)
+		Expect(cluster).NotTo(BeNil())
+		state.SetCluster(cluster)
+
+		// Create new workload cluster
+		logger.Log("Creating new workload cluster")
+
+		// We want to make sure the cluster is ready enough for us to install a new App
+		// so we wait for all control plane nodes and at least 2 workers to be ready
+		clusterReadyFns := []func(wcClient *clusterclient.Client){
+			func(wcClient *clusterclient.Client) {
+				replicas, err := state.GetFramework().GetExpectedControlPlaneReplicas(state.GetContext(), state.GetCluster().Name, state.GetCluster().GetNamespace())
+				Expect(err).NotTo(HaveOccurred())
+
+				// Only check for control plane if not a managed cluster (e.g. EKS)
+				if replicas != 0 {
+					logger.Log("Waiting for %q control plane nodes to be ready", replicas)
+					_ = wait.For(
+						wait.AreNumNodesReady(context.Background(), wcClient, 3, &cr.MatchingLabels{"node-role.kubernetes.io/control-plane": ""}),
+						wait.WithTimeout(20*time.Minute),
+						wait.WithInterval(15*time.Second),
+					)
+				}
+			},
+			func(wcClient *clusterclient.Client) {
+				logger.Log("Waiting for worker nodes to be ready")
+				_ = wait.For(
+					wait.AreNumNodesReady(context.Background(), wcClient, 2, &cr.MatchingLabels{"node-role.kubernetes.io/worker": ""}),
+					wait.WithTimeout(20*time.Minute),
+					wait.WithInterval(15*time.Second),
+				)
+			},
+			func(wcClient *clusterclient.Client) {
+				logger.Log("Waiting for all default apps to be ready")
+				defaultAppsAppName := fmt.Sprintf("%s-%s", state.GetCluster().Name, "default-apps")
+
+				Eventually(wait.IsAppDeployed(state.GetContext(), state.GetFramework().MC(), defaultAppsAppName, state.GetCluster().Organization.GetNamespace())).
+					WithTimeout(30 * time.Second).
+					WithPolling(50 * time.Millisecond).
+					Should(BeTrue())
+
+				// Wait for all default-apps apps to be deployed
+				appList := &v1alpha1.AppList{}
+				err := state.GetFramework().MC().List(state.GetContext(), appList, cr.InNamespace(state.GetCluster().Organization.GetNamespace()), cr.MatchingLabels{"giantswarm.io/managed-by": defaultAppsAppName})
+				Expect(err).NotTo(HaveOccurred())
+
+				appNamespacedNames := []types.NamespacedName{}
+				for _, app := range appList.Items {
+					appNamespacedNames = append(appNamespacedNames, types.NamespacedName{Name: app.Name, Namespace: app.Namespace})
+				}
+
+				Eventually(wait.IsAllAppDeployed(state.GetContext(), state.GetFramework().MC(), appNamespacedNames)).
+					WithTimeout(15 * time.Minute).
+					WithPolling(10 * time.Second).
+					Should(BeTrue())
+			},
+		}
+
+		cluster, err = standup.New(state.GetFramework(), false, clusterReadyFns...).Standup(cluster)
 		Expect(err).NotTo(HaveOccurred())
 		Expect(cluster).NotTo(BeNil())
 		state.SetCluster(cluster)
+
+		logger.Log("Workload cluster ready to use")
 
 		// Create app
 		app := application.New(fmt.Sprintf("%s-%s", cluster.Name, s.appName), s.appName).
@@ -155,9 +222,17 @@ func (s *suite) Run(t *testing.T, suiteName string) {
 		logger.Log("Uninstalling App %s", app.AppName)
 		err := state.GetFramework().MC().DeleteApp(state.GetContext(), *app)
 		Expect(err).NotTo(HaveOccurred())
+
+		logger.Log("Deleting workload cluster")
+		err = teardown.New(state.GetFramework()).Teardown(state.GetCluster())
+		Expect(err).NotTo(HaveOccurred())
 	})
 
 	Describe("", func() {
+		if s.afterClusterReady != nil {
+			Describe("After Cluster Ready", s.afterClusterReady)
+		}
+
 		// TODO: Exclude this if default app
 		It("Ensure app isn't already installed", func() {
 			appCR := state.GetApplication()
@@ -174,10 +249,6 @@ func (s *suite) Run(t *testing.T, suiteName string) {
 			Expect(err).ToNot(BeNil())
 			Expect(errors.IsNotFound(err)).To(BeTrue())
 		})
-
-		if s.beforeInstall != nil {
-			Describe("Before install", s.beforeInstall)
-		}
 
 		if s.isUpgrade {
 			Describe("Install previous version of app", func() {